requirements.txt
pyautoit
tika
pandas
pyodbc
pywin32
openpyxl
selenium
beautifulsoup4
requests
urllib3
pdfkit
PyPDF2
PySimpleGUI
python-dateutil
numpy
psycopg2
py4j
pillow
tkhtmlview
mysql-connector-python
python-docx